requirement.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
google-generativeai
